exercicios/19_resumo_chat_por_utilizador_sem_ecra/requirements.txt
python-dotenv>=1.0.0,<2
langchain>=1.0.0,<2
langchain-google-genai>=2.0.0,<5
langgraph>=0.2.0,<2
psycopg2-binary>=2.9.0,<3
